pyproject.toml
[tool.poetry]
name = "poppi-client"
version = "0.1.0"
description = ""
authors = ["elias-dzobo <elias.dzobo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pymongo = "^4.10.1"
faiss-cpu = "^1.8.0.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
